Update torch_dtype for model initialization
Changed torch_dtype from bfloat16 to float16 for model initialization.

+++ vlmeval/vlm/gemma.py
@@ -98,7 +98,7 @@ class Gemma3(BaseModel):
             # export VLLM_WORKER_MULTIPROC_METHOD=spawn
         else:
             self.model = Gemma3ForConditionalGeneration.from_pretrained(
-                model_path, device_map="cuda", attn_implementation="sdpa", torch_dtype=torch.bfloat16
+                model_path, device_map="cuda", attn_implementation="sdpa", torch_dtype=torch.float16
             ).eval()
             self.device = self.model.device
